# SetHoldings will check existing OnMarketOpen orders