… Algorithm.CSharp/EmitInsightNoAlphaModelAlgorithm.cs