<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- `SetHoldings` will take `OnMarketOpen` ordes into account when
determining order quantity
- Adding new regression test. Updating existing algorithms which
suffered of the issue
- Adding a performance improvement, will avoid margin and portfolio
calculations for MarketOnOpen orders that wont be able to fill

Slight performance improvement:
```
public class BasicTemplateAlgorithm : QCAlgorithm
{
    private Symbol _spy;
    public override void Initialize()
    {
        SetStartDate(2010, 1, 1);
        SetEndDate(2013, 1, 1);
        SetCash(100000);
        _spy = AddEquity("SPY").Symbol;
    }
    public override void OnData(Slice data)
    {
        SetHoldings(_spy, 1);
    }
}
```

`PR`
6.44 seconds at 46k data points per second. Processing total of 293,929 data points.
6.44 seconds at 46k data points per second. Processing total of 293,929 data points.
6.44 seconds at 46k data points per second. Processing total of 293,929 data points.
`master`
6.84 seconds at 43k data points per second. Processing total of 293,929 data points.
6.84 seconds at 43k data points per second. Processing total of 293,929 data points.
6.64 seconds at 44k data points per second. Processing total of 293,929 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3781
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
- Improve `SetHoldings` support when using multi resolution algorithms
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->